Add new Streamlit app for economic data presentation
This commit introduces a new Streamlit application with a title and subheader for the presentation of economic data. The app currently displays a title "Economic database" and a subheader "presentation of economic data". This is the initial setup and will serve as the basis for further development of the app.